fixed signed/unsigned warning

+++ test/TestAllocator.cpp
@@ -17,6 +17,6 @@ TEST_CASE("Allocator integration test")
     std::vector<double, decltype(allocator)> my_vector(allocator);
     my_vector.resize(100, 42.5);
     CHECK(resource.getFreeMemory() <= 1024 - (100 * sizeof(double)));
-    CHECK(reinterpret_cast<std::byte*>(my_vector.data()) - storage.get() < resource.getFreeMemory());
+    CHECK(reinterpret_cast<std::byte*>(my_vector.data()) - storage.get() < static_cast<int>(resource.getFreeMemory()));
     CHECK(std::all_of(begin(my_vector), end(my_vector), [](double d) { return d == 42.5; }));
 }
